--- a/.travis.yml
+++ b/.travis.yml
@@ -8,7 +8,7 @@ branches:
 os:
   - linux
 julia:
-  - 1
+  - 1.3.1
   - 1.4.1
   - nightly
 matrix: