--- a/Project.toml
+++ b/Project.toml
@@ -14,7 +14,7 @@ Combinatorics = "1.0"
 NaNMath = "0.3"
 SpecialFunctions = "0.10"
 TimerOutputs = "0.5"
-julia = "1"
+julia = "1.3"
 
 [extras]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"