--- a/src/rule_dsl.jl
+++ b/src/rule_dsl.jl
@@ -236,10 +236,25 @@ end
 node_count(atom, count; cutoff) = count + 1
 node_count(t::Term, count=0; cutoff=100) = sum(node_count(arg, count; cutoff=cutoff) for arg ∈ arguments(t))
 
-function (r::RuleSet)(term; depth=typemax(Int), applyall=false, recurse=true, threaded=false,
-                      thread_depth_cutoff=100)
+function _recurse_apply_ruleset(r::RuleSet, term; depth, recurse, applyall, threaded=false, thread_depth_cutoff=100)
     kwargs = (;applyall=applyall, recurse=recurse, threaded=threaded,
               thread_depth_cutoff=thread_depth_cutoff)
+    if threaded
+        _args = map(arguments(term)) do arg
+            if node_count(term) > thread_depth_cutoff
+                Threads.@spawn r(arg; depth=depth-1, kwargs...)
+            else
+                r(arg; depth=depth-1, kwargs..., threaded=false)
+            end
+        end
+        args = map(t -> t isa Task ? fetch(t) : t, _args)
+    else
+        args = map(t -> r(t; depth=depth-1, kwargs...), arguments(term))
+    end
+    expr = Term{symtype(term)}(operation(term), args)
+end
+
+function (r::RuleSet)(term; depth=typemax(Int), applyall=false, recurse=true, thread_kwargs...)
     rules = r.rules
     term = to_symbolic(term)
     # simplify the subexpressions
@@ -247,22 +262,10 @@ function (r::RuleSet)(term; depth=typemax(Int), applyall=false, recurse=true, th
         return term
     end
     if term isa Symbolic
-        if term isa Term && recurse
-            if threaded
-                _args = map(arguments(term)) do arg
-                    if node_count(term) > thread_depth_cutoff
-                        Threads.@spawn r(arg; depth=depth-1, kwargs...)
-                    else
-                        r(arg; depth=depth-1, kwargs..., threaded=false)
-                    end
-                end
-                args = map(t -> t isa Task ? fetch(t) : t, _args)
-            else
-                args = map(t -> r(t; depth=depth-1, kwargs...), arguments(term))
-            end
-            expr = Term{symtype(term)}(operation(term), args)
+        expr = if term isa Term && recurse
+            _recurse_apply_ruleset(r, term; depth=depth, applyall=applyall, recurse=recurse, thread_kwargs...)
         else
-            expr = term
+            term
         end
         for i in 1:length(rules)
              expr′ = try