--- a/src/rule_dsl.jl
+++ b/src/rule_dsl.jl
@@ -257,18 +257,22 @@ function _recurse_apply_ruleset(r::RuleSet, term, context; depth, recurse, apply
     kwargs = (;applyall=applyall, recurse=recurse, threaded=threaded,
               thread_subtree_cutoff=thread_subtree_cutoff)
     if threaded
-        _args = map(arguments(term)) do arg
-            if node_count(term) > thread_subtree_cutoff
-                Threads.@spawn r(arg, context; depth=depth-1, kwargs...)
-            else
-                r(arg, context; depth=depth-1, kwargs..., threaded=false)
-            end
-        end
-        args = map(t -> t isa Task ? fetch(t) : t, _args)
+        _args = _recurse_apply_ruleset_threaded_args(r, arguments(term), context, kwargs...)
     else
         args = map(t -> r(t, context; depth=depth-1, kwargs...), arguments(term))
     end
     expr = Term{symtype(term)}(operation(term), args)
+end
+
+function _recurse_apply_ruleset_threaded_args(r::RuleSet, args, context, kwargs)
+    _args = map(arguments(term)) do arg
+        if node_count(term) > thread_subtree_cutoff
+            Threads.@spawn r(arg, context; depth=depth-1, kwargs...)
+        else
+            r(arg, context; depth=depth-1, kwargs..., threaded=false)
+        end
+    end
+    map(t -> t isa Task ? fetch(t) : t, _args)
 end
 
 function (r::RuleSet)(term, context=EmptyCtx();  depth=typemax(Int), applyall=false, recurse=true, thread_kwargs...)