--- a/test/rulesets.jl
+++ b/test/rulesets.jl
@@ -77,7 +77,7 @@ end
           (1.0 / (((1 * d) / (1 + b)) * (1 / b)))) +
           ((((1 * a) + (1 * a)) / ((2.0 * (d + 1)) / 1.0)) +
            ((((d * 1) / (1 + c)) * 2.0) / ((1 / d) + (1 / c))))
-    @test simplify(ex) == simplify(ex, threaded=true, thread_depth_cutoff=3)
+    @test simplify(ex) == simplify(ex, threaded=true, thread_subtree_cutoff=3)
     @test SymbolicUtils.node_count(a + b * c / d) == 4
 end