Okay, so this "works", but it's a bit tricky to actually generate a an expression where it's worth doing. You need really big, deep terms and they can't be too 'lopsided', i.e. `(x + y) + (z + w)` is preferred to `(x + (y + (z + w)))` because it provides more opportunities for worthwhile task spawning. 

Here's something that I suspect is a bit of a 'best case' scenario. I'm using `Transducers` here to make expressions because it's `reduce` won't generate lopsided terms, unlike `foldl`. 
```julia
julia> using SymbolicUtils, Transducers

julia> begin 
           function generate_random_term(length)
               reduce(Map(_ -> rand([a,b,c,d,1,2])), 1:length, basesize=0, init=0) do x, y
                   x, y = rand(Bool) ? (x, y) : (y, x)
                   rand([*, +, /])(x, y)
               end
           end
       
           @syms a b c d
           expr1 = generate_random_term(1500)
           print("Threaded simplify \n   ")
           @btime simplify($expr1, threaded=true, fixpoint=false);
           print("Serial simplify \n   ")
           @btime simplify($expr1, threaded=false, fixpoint=false);
       end;
Threaded simplify 
     182.579 ms (3556803 allocations: 125.67 MiB)
Serial simplify 
     573.032 ms (3463980 allocations: 122.80 MiB)

julia> versioninfo()
Julia Version 1.4.1
Commit 381693d3df* (2020-04-14 17:20 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
  CPU: AMD Ryzen 5 2600 Six-Core Processor
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-8.0.1 (ORCJIT, znver1)
Environment:
  JULIA_NUM_THREADS = 6
```

This still likely needs a lot of work to make it satisfactory.