@MasonProtter 2020-05-05
ref https://github.com/JuliaSymbolics/SymbolicUtils.jl/issues/62

The threading is not enabled by default. The basic strategy is that when you do `(R::RuleSet)(term, threading=true)`, we first look at each of the arguments of `term`. If any of those arguments has an argument tree with more nodes than the keyword argument `threads_depth_cutoff=50` (needs a better name, and also tweaking for the default value), then we `@spawn` a task that applies `R` to that argument. If not, we do it in serial. Then, we map over the arguments and `fetch` any tasks that were `spawn`ed.