@MasonProtter 2020-05-05
Here's some timings on smaller expressions:
```julia
julia> begin 
           function generate_random_term(length)
               reduce(Map(_ -> rand([a,b,c,d,1,2])), 1:length, basesize=0, init=0) do x, y
                   x, y = rand(Bool) ? (x, y) : (y, x)
                   rand([*, +, /])(x, y)
               end
           end
       
           @syms a b c d
           expr1 = generate_random_term(105)
           print("Threaded simplify \n   ")
           @btime simplify($expr1, threaded=true, fixpoint=false);
           print("Serial simplify \n   ")
           @btime simplify($expr1, threaded=false, fixpoint=false);
       end;
Threaded simplify 
     25.063 ms (216607 allocations: 7.63 MiB)
Serial simplify 
     34.566 ms (214376 allocations: 7.56 MiB)
```
and 
```
julia> begin 
           function generate_random_term(length)
               reduce(Map(_ -> rand([a,b,c,d,1,2])), 1:length, basesize=0, init=0) do x, y
                   x, y = rand(Bool) ? (x, y) : (y, x)
                   rand([*, +, /])(x, y)
               end
           end
       
           @syms a b c d
           expr1 = generate_random_term(50)
           print("Threaded simplify \n   ")
           @btime simplify($expr1, threaded=true, fixpoint=false);
           print("Serial simplify \n   ")
           @btime simplify($expr1, threaded=false, fixpoint=false);
       end;
Threaded simplify 
     14.105 ms (78990 allocations: 2.73 MiB)
Serial simplify 
     14.034 ms (78468 allocations: 2.71 MiB)
```

Of course, the more time is spent in the actual rules themselves, the more we should expext this to help. As we inevitably add more sophisticated and costly rules, I'd anticipate this being more and more worth it.