@MasonProtter 2020-05-08
Here's what I'm seeing now locally:
```julia
DIFFGROUP
1-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "overhead" => 5-element BenchmarkTools.BenchmarkGroup:
          tags: []
          "ruleset" => 3-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => TrialJudgement(+0.00% => invariant)
                  "noop:Sym" => TrialJudgement(+13.21% => regression)
                  "noop:Term" => TrialJudgement(+0.27% => invariant)
          "simplify" => 3-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => TrialJudgement(-98.54% => improvement)
                  "noop:Sym" => TrialJudgement(+1.88% => invariant)
                  "noop:Term" => TrialJudgement(+2.26% => invariant)
          "acrule" => 5-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => TrialJudgement(+0.00% => invariant)
                  "noop:Sym" => TrialJudgement(+0.00% => invariant)
                  "a+2+b" => TrialJudgement(-22.63% => improvement)
                  "a+b" => TrialJudgement(-3.12% => invariant)
                  "a+2" => TrialJudgement(-23.31% => improvement)
          "simplify_no_fixp" => 3-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => TrialJudgement(-98.54% => improvement)
                  "noop:Sym" => TrialJudgement(+4.07% => invariant)
                  "noop:Term" => TrialJudgement(+1.70% => invariant)
          "rule" => 3-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => TrialJudgement(+2.18% => invariant)
                  "noop:Sym" => TrialJudgement(+1.70% => invariant)
                  "noop:Term" => TrialJudgement(+2.50% => invariant)
```