@MasonProtter 2020-05-08
Okay, I'd like to add a multi-threaded benchmark and then I think this should be good to go.