@MasonProtter 2020-05-08
I added a benchmark for simplifying a big random term `randterm:serial` and `randterm:threaded`. Here's what I'm seeing:
```julia
julia> println(j.benchmarkgroup)
1-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "overhead" => 5-element BenchmarkTools.BenchmarkGroup:
          tags: []
          "ruleset" => 3-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => TrialJudgement(+0.00% => invariant)
                  "noop:Sym" => TrialJudgement(-17.45% => improvement)
                  "noop:Term" => TrialJudgement(-10.97% => improvement)
          "simplify" => 3-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => TrialJudgement(-98.55% => improvement)
                  "noop:Sym" => TrialJudgement(+0.11% => invariant)
                  "noop:Term" => TrialJudgement(-9.31% => improvement)
          "acrule" => 5-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => TrialJudgement(+0.00% => invariant)
                  "noop:Sym" => TrialJudgement(-5.00% => invariant)
                  "a+2+b" => TrialJudgement(-15.21% => improvement)
                  "a+b" => TrialJudgement(-26.21% => improvement)
                  "a+2" => TrialJudgement(-18.69% => improvement)
          "simplify_no_fixp" => 5-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "randterm:serial" => TrialJudgement(+159.70% => regression)
                  "noop:Int" => TrialJudgement(-98.55% => improvement)
                  "randterm:thread" => TrialJudgement(+5.64% => regression)
                  "noop:Sym" => TrialJudgement(-0.39% => invariant)
                  "noop:Term" => TrialJudgement(-10.57% => improvement)
          "rule" => 3-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => TrialJudgement(-31.46% => improvement)
                  "noop:Sym" => TrialJudgement(-28.65% => improvement)
                  "noop:Term" => TrialJudgement(-19.13% => improvement)
```
It's a little mysterious to me that there's this big regression even in the serial case. I might need one additional function barrier.