@shashi 2020-05-08
aha! so these changes help the overhead for small terms even without threading!

looking great! yes those regressions are a bit mysterious.