@MasonProtter 2020-05-08
Hm, an additional function barrier around the `@spawn` ing didn't help. I'm starting to suspect the problem is the splatting of `kwargs`.