@MasonProtter 2020-05-08
Okay, I found the problem. When I refactored the recursion, I noticed that we didn't forward the `applyall` keyword argument in the recursion so I re-enabled it. However, that has *huge* performance impacts on `simplify`, so that's the difference we're seeing here.