@MasonProtter 2020-05-08
I just removed the `applyall` behaviour from the recursion, but I think it likely needs to be added back in for correctness.