@shashi 2020-05-09
OK I made the change where `(::RuleSet)` returns `nothing` if there were no changes to the expression it is rewriting.

This makes `RuleSet` behave exactly like `Rule` when given an input it doesn't care about.

Doing this means we construct new `Term`s exactly when there is a change to something.

Results are quite exciting:
```
julia> using Random, SymbolicUtils, BenchmarkTools; let
           @syms a b c d; Random.seed!(123);
           function random_term(len; atoms, funs, fallback_atom=1)
               xs = rand(atoms, len)
               while length(xs) > 1
                   xs = map(Iterators.partition(xs, 2)) do xy
                       x = xy[1]; y = get(xy, 2, fallback_atom)
                       rand(funs)(x, y)
                   end
               end
               xs[]
           end
           ex = random_term(1000; atoms=[a, b, c, d, 1, 2.0], funs=[*, /])
           @btime simplify($ex, threaded=true)
           #simplify(ex)
       end;
```

```
master:
873.904 ms (3983307 allocations: 153.98 MiB)
this branch serial:
621.633 ms (2383327 allocations: 83.35 MiB)
this branch threaded (NUM_THREADS=1):
648.166 ms (2392882 allocations: 83.56 MiB)
this branch threaded (NUM_THREADS=4):
191.065 ms (2392794 allocations: 83.56 MiB)
```